# Allow to use slices with WithObjects / WithReferences methods